Move most of pkg/util/uuid to pkg/api/util/uuid and re-export where needed
