+++ .golangci.yml
@@ -47,6 +47,8 @@ linters-settings:
 
   importas:
     alias:
+      - pkg: github.com/Azure/ARO-RP/pkg/api/util/uuid
+        alias: apiuuid
       - pkg: github.com/Azure/ARO-RP/pkg/frontend/middleware
         alias: frontendmiddleware
       - pkg: github.com/Azure/ARO-RP/pkg/metrics/statsd/cosmosdb
